Update CONTRIBUTING.md
WGs Using GitHub (wugh) is not a chartered group.  From what I could tell in the ietf-and-github mailing list archives, the ietf-and-github WG is the proper home for this work. So, I changed the outdated links by replacing 'wugh' with the updated text.  (Also see https://github.com/ietf-gitwg/using-github/issues/11 for some suggestions I made for a more general fix for this type of problem.)